Moved binning routines from adcp/tools.py to tools.py
